DAOS-12869 object: shrink layout version to 16 bits
shrink object layout version to 16 bits.

Required-githooks: true

+++ src/object/obj_ec.h
@@ -263,7 +263,7 @@ struct obj_ec_fail_info {
 };
 
 int
-obj_ec_grp_start(uint32_t layout_ver, uint64_t hash, uint32_t grp_size);
+obj_ec_grp_start(uint16_t layout_ver, uint64_t hash, uint32_t grp_size);
 
 struct obj_reasb_req;
 
@@ -370,7 +370,7 @@ is_ec_data_shard_by_tgt_off(uint32_t tgt_off, struct daos_oclass_attr *oca)
 }
 
 static inline bool
-is_ec_data_shard_by_layout_ver(uint32_t layout_ver, uint64_t dkey_hash,
+is_ec_data_shard_by_layout_ver(uint16_t layout_ver, uint64_t dkey_hash,
 			       struct daos_oclass_attr *oca, uint32_t shard)
 {
 	D_ASSERT(daos_oclass_is_ec(oca));
@@ -387,7 +387,7 @@ is_ec_parity_shard_by_tgt_off(uint32_t tgt_off, struct daos_oclass_attr *oca)
 }
 
 static inline bool
-is_ec_parity_shard_by_layout_ver(uint32_t layout_ver, uint64_t dkey_hash,
+is_ec_parity_shard_by_layout_ver(uint16_t layout_ver, uint64_t dkey_hash,
 				 struct daos_oclass_attr *oca, uint32_t shard)
 {
 	D_ASSERT(daos_oclass_is_ec(oca));
